Move table_dense_sparse test utility to Orange.widgets.tests.utils
