+++ Orange/widgets/data/tests/test_owfeaturestatistics.py
@@ -1,9 +1,9 @@
 import datetime
 import warnings
 from collections import namedtuple
-from functools import wraps, partial
+from functools import partial
 from itertools import chain
-from typing import Callable, List
+from typing import List
 
 import numpy as np
 from AnyQt.QtCore import QItemSelection, QItemSelectionRange, \
@@ -12,7 +12,7 @@ from AnyQt.QtCore import QItemSelection, QItemSelectionRange, \
 from Orange.data import Table, Domain, StringVariable, ContinuousVariable, \
     DiscreteVariable, TimeVariable
 from Orange.widgets.tests.base import WidgetTest, datasets
-from Orange.widgets.tests.utils import simulate
+from Orange.widgets.tests.utils import simulate, table_dense_sparse
 from Orange.widgets.data.owfeaturestatistics import \
     OWFeatureStatistics
 
@@ -173,18 +173,6 @@ def make_table(attributes, target=None, metas=None):
         Domain(attribute_vars, class_vars=target_vars, metas=meta_vars),
         X=attribute_vals, Y=target_vals, metas=meta_vals,
     )
-
-
-def table_dense_sparse(test_case):
-    # type: (Callable) -> Callable
-    """Run a single test case on both dense and sparse Orange tables."""
-
-    @wraps(test_case)
-    def _wrapper(self):
-        test_case(self, lambda table: table.to_dense())
-        test_case(self, lambda table: table.to_sparse())
-
-    return _wrapper
 
 
 class TestVariousDataSets(WidgetTest):
